+++ src/addons/godot_devtool/runtime_bridge.gd
@@ -9,6 +9,10 @@ func _ready() -> void:
 	_client.ready()
 	set_process(true)
 
+func _exit_tree() -> void:
+	set_process(false)
+	_client.exit_tree()
+
 func _input(event: InputEvent) -> void:
 	_client.input(event)